Move the Epoch type to a separate module
- This should help with avoiding a dependency cycle in upcoming changes
