+++ libs/wire-api/src/Wire/API/Conversation/Protocol.hs
@@ -39,8 +39,8 @@ import Data.Schema
 import Imports
 import Wire.API.Arbitrary
 import Wire.API.Conversation.Action.Tag
+import Wire.API.MLS.Epoch
 import Wire.API.MLS.Group
-import Wire.API.MLS.Message
 
 data ProtocolTag = ProtocolProteusTag | ProtocolMLSTag
   deriving stock (Eq, Show, Enum, Bounded, Generic)
