+++ libs/wire-api/src/Wire/API/MLS/Message.hs
@@ -20,8 +20,7 @@
 -- with this program. If not, see <https://www.gnu.org/licenses/>.
 
 module Wire.API.MLS.Message
-  ( Epoch (..),
-    Message (..),
+  ( Message (..),
     WireFormatTag (..),
     SWireFormatTag (..),
     SomeMessage (..),
@@ -35,23 +34,16 @@ module Wire.API.MLS.Message
 where
 
 import Data.Binary
-import Data.Schema
 import Data.Singletons.TH
 import qualified Data.Swagger as S
 import Imports
-import Wire.API.Arbitrary
 import Wire.API.MLS.Commit
+import Wire.API.MLS.Epoch
 import Wire.API.MLS.Group
 import Wire.API.MLS.KeyPackage
 import Wire.API.MLS.Proposal
 import Wire.API.MLS.Serialisation
 
-newtype Epoch = Epoch {epochNumber :: Word64}
-  deriving stock (Eq, Show)
-  deriving newtype (Arbitrary, Enum, ToSchema)
-
-instance ParseMLS Epoch where
-  parseMLS = Epoch <$> parseMLS
 
 data WireFormatTag = MLSPlainText | MLSCipherText
   deriving (Bounded, Enum, Eq, Show)
